Moving actual filter controllers to correct places
